+++ lib/dhtml/document.rb
@@ -130,17 +130,15 @@ module DHTML
     #
     # @!attribute [String] tag
     # @!attribute [Hash] attributes
-    # @return [Integer] Number of bytes written to the stream.
+    # @return [void]
     # @since 0.1.0
     def write_html_element(tag, attributes = {})
-      length = document.length
-
       document << '<'
       document << tag
       document << " #{html_attributes(attributes)}" unless attributes.empty?
       document << '>'
 
-      document.length - length
+      nil
     end
 
     # Write a tag to the HTML document.
@@ -158,9 +156,6 @@ module DHTML
       tag = tag.to_s
       tag = tag[1..-1] if tag[0] == '_'
 
-      # Used to calculate the number of bytes written for the return value.
-      length = document.length
-
       # Opening tag with its HTML attributes - e.g. <div id="main">
       write_html_element(tag, attributes)
 
@@ -173,8 +168,7 @@ module DHTML
       # Close the tag when necessary.
       document.write("</#{tag}>") if content.is_a?(String) || block_given? || !void?(tag.to_sym)
 
-      # Total number of bytes written by this method.
-      document.length - length
+      nil
     end
   end
 end