+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [0.1.5] - 2021-05-11
+### Removed
+- Bytes written is no longer returned since `StringIO` is not fully implemented by Opal.
+
 ## [0.1.4] - 2021-05-11
 ### Added
 - `pretty_html` method that formats the document in a way presentable to a human.
@@ -17,6 +21,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - Versions prior to this were yanked. The gems were packaged with absolute paths by mistake.
 
 [Unreleased]: https://github.com/hi5dev/dhtml/compare/v0.1.4...HEAD
+[0.1.5]: https://github.com/hi5dev/dhtml/compare/v0.1.4..v0.1.5
 [0.1.4]: https://github.com/hi5dev/dhtml/compare/v0.1.3..v0.1.4
 [0.1.3]: https://github.com/hi5dev/dhtml/releases/tag/v0.1.3